# AIM-OpenAPI.yaml
openapi: 3.0.3
...
info:
  title: Armory Inventory Management APIs
  description: APIs for managing AIM (Armory Inventory Management)
  version: 1.0.0
...
paths:
...
  /departments:
    get:
      summary: Get all departments
      operationId: getAllDepartments
      tags:
        - Departments
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: A list of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

    post:
      summary: Create a new department
      operationId: createDepartment
      tags:
        - Departments
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentDto'
      responses:
        '201':
          description: Department created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDto'
        '400':
          description: Bad request - Missing or invalid data
        '409':
          description: Department with same name already present.
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /departments/{departmentId}:
    get:
      summary: Get department by departmentId
      operationId: getDepartmentById
      tags:
        - Departments
      parameters:
        - name: departmentId
          in: path
          required: true
          description: ID of the department to fetch
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDto'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Department not found"

    put:
      summary: Update department by department Id
      operationId: updateDepartment
      tags:
        - Departments
      parameters:
        - name: departmentId
          in: path
          required: true
          description: ID of the department to update
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentDto'
      responses:
        '200':
          description: Department updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDto'
        '404':
          description: Department data not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Department not found"
        '400':
          description: Bad request
        '500':
          description: Internal server error

    delete:
      summary: Delete department by ID
      operationId: deleteDepartment
      tags:
        - Departments
      parameters:
        - name: departmentId
          in: path
          required: true
          description: ID of the department to delete
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: Department deleted successfully
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Department not found"
        '500':
          description: Internal server error
  /departments/name/{departmentName}:
    get:
      summary: Get department by name
      operationId: getDepartmentByName
      tags:
        - Departments
      parameters:
        - name: departmentName
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDto'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Department not found"
#
#  /departments/organization/{organizationId}:
#    get:
#      summary: Get all departments by organization ID
#      operationId: getDepartmentsByOrganizationId
#      tags:
#        - Departments
#      parameters:
#        - name: organizationId
#          in: path
#          required: true
#          schema:
#            type: integer
#            format: int64
#      responses:
#        '200':
#          description: A list of departments belonging to the organization
#          content:
#            application/json:
#              schema:
#                type: array
#                items:
#                  $ref: '#/components/schemas/DepartmentDto'
#        '404':
#          description: Department not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Department not found"
#
#==============================================Manufacturer APIs====================================================

  # Get all manufacturers
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    X-Tenant-ID:
      name: X-Tenant-ID
      in: header
      description: Tenant identifier for multi-tenancy.
      required: true
      schema:
        type: string
#    Authorization:
#      name: Authorization
#      in: header
#      description: "Bearer token for authentication. Format: `Bearer <token>`"
#      required: true
#      schema:
#        type: string
  schemas:
...
    OrganizationDto:
      type: object
      properties:
        organizationId:
          type: integer
          format: int64
          description: The unique identifier of the organization.
        organizationName:
          type: string
          description: The unique Name of the organization.
        organizationDetails:
          type: object
          additionalProperties: true  # Allows storing arbitrary JSON data
          description: A JSON object containing details such as organization name, username, email, and phone number
          example:
            email: "admin@weaponvault.com"
            phone_no: "+1234567890"
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the organization was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the organization was last updated.
        status:
          type: string
          description: The status of the organization (active, inactive, etc.)
#        departments: # One-to-many relationship to departments
#          type: array
#          items:
#            $ref: '#/components/schemas/DepartmentDto'
      required:
        - organizationName
        - organizationDetails

    #Department Schema
    DepartmentDto:
      type: object
      properties:
        departmentId:
          type: integer
          format: int64
          description: The unique identifier of the department.
        departmentName:
          type: string
          description: The name of the department.
        departmentDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing details about the department.
          example:
            department_code: "HR001"
            department_head: "John Doe"
            description: "Handles all human resources tasks."
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the department was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the department was last updated.
#        organization:  # Many-to-one relationship back to the organization
#          $ref: '#/components/schemas/OrganizationDto'
      required:
        - departmentName
        - departmentDetails


    # Manufacturer Schemas
...