# src/main/java/com/armory/WeaponVault/organizationApi/DepartmentsApiDelegate.java
package com.armory.WeaponVault.organizationApi;

import com.armory.WeaponVault.entityDTO.DepartmentDto;
import com.armory.WeaponVault.entityDTO.GetAllOrganizations500Response;
import com.armory.WeaponVault.entityDTO.GetDepartmentById404Response;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

/**
 * A delegate to be called by the {@link DepartmentsApiController}}.
 * Implement this interface with a {@link org.springframework.stereotype.Service} annotated class.
 */
@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-04-29T14:12:51.910033071+05:30[Asia/Kolkata]", comments = "Generator version: 7.11.0")
public interface DepartmentsApiDelegate {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /departments : Create a new department
     *
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @param departmentDto  (required)
     * @return Department created successfully (status code 201)
     *         or Bad request - Missing or invalid data (status code 400)
     *         or Department with same name already present. (status code 409)
     *         or Internal server error (status code 500)
     * @see DepartmentsApi#createDepartment
     */
    default ResponseEntity<DepartmentDto> createDepartment(String xTenantID,
        DepartmentDto departmentDto) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"departmentName\" : \"departmentName\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"departmentDetails\" : { \"department_code\" : \"HR001\", \"department_head\" : \"John Doe\", \"description\" : \"Handles all human resources tasks.\" }, \"departmentId\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"Internal server error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * DELETE /departments/{departmentId} : Delete department by ID
     *
     * @param departmentId ID of the department to delete (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return Department deleted successfully (status code 200)
     *         or Department not found (status code 404)
     *         or Internal server error (status code 500)
     * @see DepartmentsApi#deleteDepartment
     */
    default ResponseEntity<Void> deleteDepartment(Long departmentId,
        String xTenantID) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"Department not found\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * GET /departments : Get all departments
     *
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return A list of departments (status code 200)
     *         or Internal server error (status code 500)
     * @see DepartmentsApi#getAllDepartments
     */
    default ResponseEntity<List<DepartmentDto>> getAllDepartments(String xTenantID) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"departmentName\" : \"departmentName\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"departmentDetails\" : { \"department_code\" : \"HR001\", \"department_head\" : \"John Doe\", \"description\" : \"Handles all human resources tasks.\" }, \"departmentId\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"departmentName\" : \"departmentName\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"departmentDetails\" : { \"department_code\" : \"HR001\", \"department_head\" : \"John Doe\", \"description\" : \"Handles all human resources tasks.\" }, \"departmentId\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"Internal server error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * GET /departments/{departmentId} : Get department by departmentId
     *
     * @param departmentId ID of the department to fetch (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return Department details (status code 200)
     *         or Department not found (status code 404)
     * @see DepartmentsApi#getDepartmentById
     */
    default ResponseEntity<DepartmentDto> getDepartmentById(Long departmentId,
        String xTenantID) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"departmentName\" : \"departmentName\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"departmentDetails\" : { \"department_code\" : \"HR001\", \"department_head\" : \"John Doe\", \"description\" : \"Handles all human resources tasks.\" }, \"departmentId\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"Department not found\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * GET /departments/name/{departmentName} : Get department by name
     *
     * @param departmentName  (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @return Department details (status code 200)
     *         or Department not found (status code 404)
     * @see DepartmentsApi#getDepartmentByName
     */
    default ResponseEntity<DepartmentDto> getDepartmentByName(String departmentName,
        String xTenantID) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"departmentName\" : \"departmentName\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"departmentDetails\" : { \"department_code\" : \"HR001\", \"department_head\" : \"John Doe\", \"description\" : \"Handles all human resources tasks.\" }, \"departmentId\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"Department not found\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * PUT /departments/{departmentId} : Update department by department Id
     *
     * @param departmentId ID of the department to update (required)
     * @param xTenantID Tenant identifier for multi-tenancy. (required)
     * @param departmentDto  (required)
     * @return Department updated successfully (status code 200)
     *         or Department data not found (status code 404)
     *         or Bad request (status code 400)
     *         or Internal server error (status code 500)
     * @see DepartmentsApi#updateDepartment
     */
    default ResponseEntity<DepartmentDto> updateDepartment(Long departmentId,
        String xTenantID,
        DepartmentDto departmentDto) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"departmentName\" : \"departmentName\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"departmentDetails\" : { \"department_code\" : \"HR001\", \"department_head\" : \"John Doe\", \"description\" : \"Handles all human resources tasks.\" }, \"departmentId\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"Department not found\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}